marketSpace/requirements.txt
asgiref==3.7.2
cachetools==4.2.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0
django-allauth==0.60.0
django-cors-headers==4.3.1
django-googledrive-storage==1.6.0
django-js-asset==2.2.0
django-mptt==0.15.0
djangorestframework==3.14.0
google-api-core==2.10.2
google-api-python-client==2.111.0
google-auth==1.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
